Split DateTime component out of input.js
and fix react and eslint errors.
